% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedToDataframeCopyNumber.R
\name{BedToDataframeCopyNumber}
\alias{BedToDataframeCopyNumber}
\title{BedToDataframeCopyNumber}
\usage{
BedToDataframeCopyNumber(bed)
}
\arguments{
\item{bed}{filepath to the ".bed" file}
}
\description{
BedToDataframeCopyNumber
}
\author{
Ulrich Asemann
}
