% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caretLogLoss.R
\name{caretLogLoss}
\alias{caretLogLoss}
\title{caretLogLoss Function}
\usage{
caretLogLoss(data, lev = NULL, model = NULL)
}
\arguments{
\item{model}{}
}
\value{
Returns the caretLogLoss.
}
\description{
The selfwritten caretLogLoss function for the package.
}
\author{
Ulrich Asemann
}
