% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPartitioning.R
\name{DataPartitioning}
\alias{DataPartitioning}
\title{Data Partitioning}
\usage{
DataPartitioning(
  phenotypeIDs,
  dataIDs,
  partitioning = 0.8,
  numPartitions = 100,
  type = "training",
  seed = 123
)
}
\arguments{
\item{phenotypeIDs}{A table holding the clinical IDs with a variable indicating,
if the disease occurred or not.}

\item{dataIDs}{A table holding all the clinical IDs with the respective IDs
for each modality.}

\item{partitioning}{A value, that signals what percentage of IDs are used for
training in the feature selection.}

\item{numPartitions}{The number of partitions that are build for feature selection}

\item{type}{This variable signals, if the training or testing set of the
feature selection is returned. Otherwise the type can be set to "modeltraining"
to return the set of IDs that are used for modeltraining.}

\item{seed}{The possibility to change the seed, for different results in the
part of the ID shuffling for feature selection.}
}
\value{
The return of the function can be three different outcomes depending
on the selected type. The return is always a list containing the IDs for feature
selection or modeltraining.
}
\description{
Builds a data partition for the provided sets of IDs.
}
\author{
Ulrich Asemann
}
