% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{EastCirclePine}
\alias{EastCirclePine}
\title{Volume, height, and diameter of 300 pine trees in East Circle Jammu}
\format{
An object of class \code{data.frame} with 300 rows and 3 columns.
}
\usage{
EastCirclePine
}
\description{
A data frame consisting of diameter (cm), height (m), 
             and volume (m3) for 300 pine trees in East Circle, Jammu,
             Jammu and Kashmir Forest Department.
}
\keyword{datasets}
