% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature Selection}
\usage{
feature_selection()
}
\value{
The result of another function, depending on the chosen modailty
}
\description{
A function that calls the right feature selection function,
depending on the modality that is chosen by the user
}
