% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Feature Selection}
\usage{
FeatureSelection()
}
\value{
The result of another function, depending on the chosen modailty
}
\description{
A function that calls the right feature selection function,
depending on the modality that is chosen by the user
}
\author{
Ulrich Asemann
}
