% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methylomics.R
\name{fs_methylomics}
\alias{fs_methylomics}
\title{Title}
\usage{
fs_methylomics(
  train_IDs,
  test_IDs,
  data_IDs,
  phenotype_IDs,
  methylomics_data,
  low_lev_path,
  seed = 123
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
