% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_proteomics.R
\name{fs_proteomics}
\alias{fs_proteomics}
\title{Title}
\usage{
fs_proteomics(
  train_IDs,
  test_IDs,
  data_IDs,
  phenotype_IDs,
  proteomics_data,
  low_lev_path,
  somamaer_info_edited,
  seed = 123
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
