% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_transcriptomics.R
\name{fs_transcriptomics}
\alias{fs_transcriptomics}
\title{Feature Selection for Transcriptomics}
\usage{
fs_transcriptomics(
  train_IDs,
  test_IDs,
  data_IDs,
  phenotype_IDs,
  transcriptomics_data,
  gene_anotation,
  low_lev_path,
  seed = 123
)
}
\arguments{
\item{seed}{}
}
\description{
Feature Selection for Transcriptomics
}
