% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FsClinical.R
\name{FsClinical}
\alias{FsClinical}
\title{Feature Selection Clinical Data}
\usage{
FsClinical(trainIDs, testIDs, dataIDs, phenotypeIDs, clinicalData)
}
\arguments{
\item{trainIDs}{Set of training IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{testIDs}{Set of testing IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{dataIDs}{A table holding all the clinical IDs with the respective IDs
for each modality.}

\item{phenotypeIDs}{A table holding the clinical IDs with a variable indicating,
if the disease occurred or not.}

\item{clinicalData}{A table holding the clinical data.}
}
\value{
Returns the clinical data table only with the significant selected features.
}
\description{
The function for the feature selection for the clinical data.
}
\author{
Ulrich Asemann
}
