% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FsGenomics.R
\name{FsGenomics}
\alias{FsGenomics}
\title{Feature Selection Genomics Data}
\usage{
FsGenomics(lowestLevelPathways, genomicsAnnotation, seed = 123)
}
\arguments{
\item{lowestLevelPathways}{The lowest level pathways for the genomics.}

\item{genomicsAnnotation}{An annotation file that is needed for the function to work}

\item{seed}{A seed that can be set by the user to achieve the same results}
}
\value{
A text file will be saved in the environment holding the results.
}
\description{
The function for the feature selection for the genomics data.
}
\author{
Ulrich Asemann
}
