% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FsMetabolomics.R
\name{FsMetabolomics}
\alias{FsMetabolomics}
\title{Feature Selection Metabolomics Data}
\usage{
FsMetabolomics(
  trainIDs,
  testIDs,
  dataIDs,
  phenotypeIDs,
  metabolomicsData,
  pathwayList,
  lowestLevelPathways,
  annotationFile,
  seed = 123
)
}
\arguments{
\item{trainIDs}{Set of training IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{testIDs}{Set of testing IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{dataIDs}{A table holding all the clinical IDs with the respective IDs
for each modality.}

\item{phenotypeIDs}{A table holding the clinical IDs with a variable indicating,
if the disease occurred or not.}

\item{metabolomicsData}{A table holding the data for the metabolomics.}

\item{pathwayList}{A list containing the pathways for the metabolomics.}

\item{lowestLevelPathways}{The lowest level pathways.}

\item{annotationFile}{An annotation file containing additional information for the metabolomics.}

\item{seed}{The possibility to change the seed for the function.}
}
\value{
Returns the metabolomics data table with the selected features.
}
\description{
The function for the feature selection for metabolomics data.
}
\author{
Ulrich Asemann
}
