% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{Loading necessary data}
\usage{
load(
  clinical = FALSE,
  metabolomics = FALSE,
  methylomics = FALSE,
  protemoics = FALSE,
  transcriptomics = FALSE
)
}
\arguments{
\item{transcriptomics}{}
}
\value{
Loads the selected data to the global environment.
}
\description{
Loading all necessary data for building the package. The values
for the parameters can be TRUE or FALSE, if the respective data for the
modality should be oaded or not.
}
\author{
Ulrich Asemann
}
