% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLoss.R
\name{LogLoss}
\alias{LogLoss}
\title{LogLoss Function}
\usage{
LogLoss(pred, true, eps = 1e-15, weights = NULL)
}
\arguments{
\item{weights}{}
}
\value{
Returns the LogLoss.
}
\description{
The self written LogLoss function for the package.
}
\author{
Ulrich Asemann
}
