% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeOverlapSet.R
\name{MakeOverlapSet}
\alias{MakeOverlapSet}
\title{MakeOverlapSet}
\usage{
MakeOverlapSet(dataList, IDTable, phenotypeIDs)
}
\arguments{
\item{phenotypeIDs}{}
}
\description{
MakeOverlapSet
}
\author{
Ulrich Asemann
}
