% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectModeltrainingIDs.R
\name{SelectModeltrainingIDs}
\alias{SelectModeltrainingIDs}
\title{SelectModeltrainingIDs}
\usage{
SelectModeltrainingIDs(dataFrame, modeltrainingIDs, dataIDs, modality)
}
\arguments{
\item{modality}{}
}
\value{
the modality data frame with only the modeltraining IDs
}
\description{
SelectModeltrainingIDs
}
\author{
Ulrich Asemann
}
