% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardise.R
\name{Standardise}
\alias{Standardise}
\title{Standardise}
\usage{
Standardise(dataFrame)
}
\arguments{
\item{dataFrame}{to be standardised}
}
\value{
a standardised data frame
}
\description{
Standardises a data frame
}
\author{
Ulrich Asemann
}
