% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partitioning.R
\name{data_partitioning}
\alias{data_partitioning}
\title{Data Partitioning Function}
\usage{
data_partitioning(
  phenotype_IDs,
  data,
  partitioning = 0.8,
  type = "training",
  seed = 123
)
}
\arguments{
\item{type}{}
}
\description{
Returns a list with all training IDs and feature selection IDs,
that are for testing or training
}
\author{
Ulrich Asemann
}
