% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methylomics.R
\name{fs_methylomics}
\alias{fs_methylomics}
\title{Feature Selection Methylomics Data}
\usage{
fs_methylomics(
  train_IDs,
  test_IDs,
  data_IDs,
  phenotype_IDs,
  methylomics_data,
  low_lev_path,
  seed = 123
)
}
\arguments{
\item{train_IDs}{Set of training IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{test_IDs}{Set of testing IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{data_IDs}{A table holding all the clinical IDs with the respective IDs
for each modality.}

\item{phenotype_IDs}{A table holding the clinical IDs with a variable indicating,
if the disease occurred or not.}

\item{methylomics_data}{A table holding the data for the methylomics.}

\item{low_lev_path}{The lowest level pathways.}

\item{seed}{The possibility to change the seed for the function.}
}
\value{
Returns the methylomics data table with the selected features.
}
\description{
The function for the feature selection for methylomics data.
}
\author{
Ulrich Asemann
}
