% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_transcriptomics.R
\name{fs_transcriptomics}
\alias{fs_transcriptomics}
\title{Feature Selection for transcriptomics}
\usage{
fs_transcriptomics(
  train_IDs,
  test_IDs,
  data_IDs,
  phenotype_IDs,
  transcriptomics_data,
  gene_anotation = NULL
)
}
\arguments{
\item{transcriptomics_data}{}
}
\description{
Feature Selection for transcriptomics
}
