% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FsTranscriptomics.R
\name{fs_transcriptomics}
\alias{fs_transcriptomics}
\title{Feature Selection Transcriptomics Data}
\usage{
fs_transcriptomics(
  trainIDs,
  testIDs,
  dataIDs,
  phenotypeIDs,
  transcriptomicsData,
  geneAnotation,
  lowestLevelPathways,
  seed = 123
)
}
\arguments{
\item{trainIDs}{Set of training IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{testIDs}{Set of testing IDs from the \code{data_partitioning()} function
for the feature selection.}

\item{dataIDs}{A table holding all the clinical IDs with the respective IDs
for each modality.}

\item{phenotypeIDs}{A table holding the clinical IDs with a variable indicating,
if the disease occurred or not.}

\item{transcriptomicsData}{A table holding the data for the transcriptomics.}

\item{geneAnotation}{An annotation file for the transcriptomics data.}

\item{lowestLevelPathways}{The lowest level pathways}

\item{seed}{The possibility to change the seed for the function.}
}
\value{
Returns the transcriptomics data table with the selected features.
}
\description{
The function for the feature selection for transcriptomics data.
}
\author{
Ulrich Asemann
}
