% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_transcriptomics.R
\name{fs_transcriptomics}
\alias{fs_transcriptomics}
\title{Feature Selection for Transcriptomics}
\usage{
fs_transcriptomics(IDs, data_IDs, transcriptomics_data)
}
\arguments{
\item{transcriptomics_data}{}
}
\description{
Feature Selection for Transcriptomics
}
