% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{load()}
\usage{
load(
  clinical = FALSE,
  metabolomics = FALSE,
  methylomics = FALSE,
  protemoics = FALSE,
  transcriptomics = FALSE
)
}
\description{
Loading all necessary data, to build the package
}
\examples{
load()
}
